fix building by moving components out of pages
