+++ components/Calculator/calculator.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/globals.scss";
+@import "../../styles/globals.scss";
 $color-numbers: #F6ECFF;
 $color-secondaryOperands: #F1E0FD;
 $color-operands-1: #F7D261;
