+++ components/Calculator/functions.js
