+++ components/Footer/Footer.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames/bind';
+import classNames from 'classnames';
 import styles from './footer.module.scss';
 
 export default function Footer({ text="Design created by", link, author, className }) {
