+++ components/HomePage/home.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/globals.scss";
+@import "../../styles/globals.scss";
 
 .main {
   @include page;
