+++ components/MediaCard/mediaCard.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/globals.scss";
+@import "../../styles/globals.scss";
 
 .card {
     margin: 1rem;
