+++ pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Home from './components/HomePage/Home';
+import Home from '../components/HomePage/Home';
 
 export default function index() {
   return <Home />;
